+++ src/config.php
@@ -3,4 +3,5 @@
 return [
     // The root path of your generated JSON language files
     'path' => lang_path(''),
+    'apps' => null,
 ];